add changes and merge monte-carlo scripts
