+++ scripts/monte_carlo.py
@@ -77,6 +77,17 @@ from solve_network import *
 logger = logging.getLogger(__name__)
 
 
+def wildcard_creator(config, method=None):
+    """
+    Creates wildcard for monte-carlo simulations.
+    """
+    if method=="global_sensitivity":
+        return [f"g{i}" for i in range(config["monte_carlo"]["options"]["samples"])]
+    
+    elif method=="any_chance_store_test":
+        return [f"a{i}" for i in range(len(config["electricity"]["extendable_carriers"]["Store"]))]
+
+
 def monte_carlo_sampling_pydoe2(
     N_FEATURES,
     SAMPLES,
@@ -158,6 +169,33 @@ def monte_carlo_sampling_scipy(
     return lh
 
 
+def single_best_in_worst_list(worst_list, best_list):
+    """
+    Return list with single best value per list of worst values
+    
+    Input:
+    ------
+    worst_list: 1D array
+    best_list: 1D array
+    
+    Output:
+    -------
+    new_list: 2D array
+    
+    Example:
+    --------
+    >>> single_best_in_worst_list([1, 1, 1], [4, 4, 4])
+    [[4, 1, 1], [1, 4, 1], [1, 1, 4]]
+    """
+    new_list = []
+    for i in range(len(worst_list)):
+        l = worst_list.copy()
+        l[i] = best_list[i]
+        new_list.append(l)
+    
+    return new_list
+
+
 if __name__ == "__main__":
     if "snakemake" not in globals():
         from _helpers import mock_snakemake
@@ -173,72 +211,94 @@ if __name__ == "__main__":
         )
     configure_logging(snakemake)
     config = snakemake.config
+    n = pypsa.Network(snakemake.input[0])
 
-    ### SCENARIO INPUTS
-    ###
-    MONTE_CARLO_PYPSA_FEATURES = {
+    PYPSA_FEATURES = {
         k: v for k, v in config["monte_carlo"]["pypsa_standard"].items() if v
     }  # removes key value pairs with empty value e.g. []
-    MONTE_CARLO_OPTIONS = config["monte_carlo"]["options"]
-    L_BOUNDS = [item[0] for item in MONTE_CARLO_PYPSA_FEATURES.values()]
-    U_BOUNDS = [item[1] for item in MONTE_CARLO_PYPSA_FEATURES.values()]
+    OPTIONS = config["monte_carlo"]["options"]
+    L_BOUNDS = [item[0] for item in PYPSA_FEATURES.values()]
+    U_BOUNDS = [item[1] for item in PYPSA_FEATURES.values()]
     N_FEATURES = len(
-        MONTE_CARLO_PYPSA_FEATURES
+        PYPSA_FEATURES
     )  # only counts features when specified in config
-    SAMPLES = MONTE_CARLO_OPTIONS.get(
-        "samples"
-    )  # TODO: What is the optimal sampling? Fabian Neumann answered that in "Broad ranges" paper
-    SAMPLING_STRATEGY = MONTE_CARLO_OPTIONS.get("sampling_strategy")
 
-    ### SCENARIO CREATION / SAMPLING STRATEGY
+    ### SCENARIO CREATION
     ###
-    if SAMPLING_STRATEGY == "pydoe2":
-        lh = monte_carlo_sampling_pydoe2(
-            N_FEATURES,
-            SAMPLES,
-            criterion=None,
-            iteration=None,
-            random_state=42,
-            correlation_matrix=None,
-        )
-    if SAMPLING_STRATEGY == "scipy":
-        lh = monte_carlo_sampling_scipy(
-            N_FEATURES,
-            SAMPLES,
-            centered=False,
-            strength=2,
-            optimization=None,
-            seed=42,
-        )
-    if SAMPLING_STRATEGY == "chaospy":
-        lh = monte_carlo_sampling_chaospy(
-            N_FEATURES,
-            SAMPLES,
-            rule="latin_hypercube",
-            seed=42,
-        )
-    lh_scaled = qmc.scale(lh, L_BOUNDS, U_BOUNDS)
+    if OPTIONS.get("method")=="global_sensitivity":
+        SAMPLES = OPTIONS.get(
+            "samples"
+        )  # TODO: What is the optimal sampling? Fabian Neumann answered that in "Broad ranges" paper    
+        SAMPLING_STRATEGY = OPTIONS.get("sampling_strategy")
 
-    ### MONTE-CARLO MODIFICATIONS
+        if SAMPLING_STRATEGY == "pydoe2":
+            lh = monte_carlo_sampling_pydoe2(
+                N_FEATURES,
+                SAMPLES,
+                criterion=None,
+                iteration=None,
+                random_state=42,
+                correlation_matrix=None,
+            )
+        if SAMPLING_STRATEGY == "scipy":
+            lh = monte_carlo_sampling_scipy(
+                N_FEATURES,
+                SAMPLES,
+                centered=False,
+                strength=2,
+                optimization=None,
+                seed=42,
+            )
+        if SAMPLING_STRATEGY == "chaospy":
+            lh = monte_carlo_sampling_chaospy(
+                N_FEATURES,
+                SAMPLES,
+                rule="latin_hypercube",
+                seed=42,
+            )
+        scenarios = qmc.scale(lh, L_BOUNDS, U_BOUNDS)
+
+    elif OPTIONS.get("method")=="any_chance_store_test":
+        carrier_no = len(n.stores.carrier.unique())
+        worst_list = L_BOUNDS * carrier_no
+        best_list = U_BOUNDS * carrier_no
+        scenarios = single_best_in_worst_list(worst_list, best_list)
+
+    ### NETWORK MODIFICATION
     ###
-    n = pypsa.Network(snakemake.input[0])
     unc_wildcards = snakemake.wildcards[-1]
     i = int(unc_wildcards[1:])
     j = 0
-    for k, v in MONTE_CARLO_PYPSA_FEATURES.items():
-        # this loop sets in one scenario each "i" feature assumption
-        # k is the config input key "loads_t.p_set"
-        # v is the lower and upper bound [0.8,1.3], that was used for lh_scaled
-        # i, j interation number to pick values of experimental setup
-        # Example: n.loads_t.p_set = network.loads_t.p_set = .loads_t.p_set * lh_scaled[0,0]
-        exec(f"n.{k} = n.{k} * {lh_scaled[i,j]}")
-        logger.info(f"Scaled n.{k} by factor {lh_scaled[i,j]} in the {i} scenario")
-        j = j + 1
+
+    if OPTIONS.get("method")=="global_sensitivity":
+        for k, v in PYPSA_FEATURES.items():
+            # this loop sets in one scenario each "i" feature assumption
+            # k is the config input key "loads_t.p_set"
+            # v is the lower and upper bound [0.8,1.3], that was used for scenarios
+            # i, j interation number to pick values of experimental setup
+            # Example: n.loads_t.p_set = n.loads_t.p_set * scenarios[0,0]
+            exec(f"n.{k} = n.{k} * {scenarios[i,j]}")
+            logger.info(f"Scaled n.{k} by factor {scenarios[i,j]} in the {i} scenario")
+            j = j + 1
+
+    if OPTIONS.get("method")=="any_chance_store_test":
+        for k, _ in PYPSA_FEATURES.items():
+            type = k.split(".")[0] # "stores", "generators", ...
+            feature = k.split(".")[1] # "capital_cost", "efficiency", ...
+
+        if type == "stores":
+            # scales the whole storage-chain
+            carrier_list = n.stores.carrier.unique()
+            for c in carrier_list:
+                n.stores.loc[n.stores.carrier==c, feature] *= scenarios[i][j]
+                n.links.loc[n.links.carrier.str.contains("H2"), feature] *= scenarios[i][j] 
+                logger.info(f"Scaled {feature} for carrier={c} of store and links by factor {scenarios[i][j]} in the {i} scenario")
+                j += 1
 
     ### EXPORT AND METADATA
-    #
-    latin_hypercube_dict = (
-        pd.DataFrame(lh_scaled).rename_axis("Nruns").add_suffix("_feature")
+    ###
+    scenario_dict = (
+        pd.DataFrame(scenarios).rename_axis("iteration").add_suffix("_feature")
     ).to_dict()
-    n.meta.update(latin_hypercube_dict)
+    n.meta.update(scenario_dict)
     n.export_to_netcdf(snakemake.output[0])
