+++ test/config.monte_carlo.yaml
@@ -7,5 +7,4 @@ retrieve_databundle: # required to be "false" for nice CI test output
   show_progress: false
 
 monte_carlo:
-  options:
-    add_to_snakefile: true
+  add_to_snakefile: true
